solved bugs

+++ 33-gluPerspectiveTemplate/Template/Template.cpp
@@ -297,7 +297,7 @@ void resize(int width, int height)
 
 	glMatrixMode(GL_PROJECTION);
 	glLoadIdentity();
-	gluPerspective(45.0f,width/height,0.1f,100.0f);
+	gluPerspective(45.0f,(GLfloat)width/ (GLfloat)height,0.1f,100.0f);
 }
 
 void uninitialize(void)
